% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{detect_language}
\alias{detect_language}
\title{Automatically detect the language of a given string}
\usage{
detect_language(str)
}
\arguments{
\item{str}{The string you want to detect the language}
}
\description{
You don't know in which language is a string written? Detect it using this function.
}
\examples{
\dontrun{
detect_language("what language is this?")
}
}
