% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{get_languages}
\alias{get_languages}
\title{Get a list of supported languages for your working instance}
\usage{
get_languages()
}
\description{
Run this function and you will be able to know the supported languages for this
instance and also their ISO-639 code, needed to set the from and to languages
when using \code{translate()} function.
}
\examples{
\dontrun{
get_languages()
}
}
