% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{import_config}
\alias{import_config}
\title{Import configuration for an instance for a file}
\usage{
import_config(file)
}
\arguments{
\item{file}{.json file generated by \code{set_config()} with the needed information}
}
\description{
Read the file that contains your instance and the API key for it (if needed),
instead of having to execute the \code{set_config()} wizard again.
}
\examples{
\dontrun{
import_config("my-custom-instance.json")
}
}
