% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_config}
\alias{set_config}
\title{Set the LibreTranslate instance you want to use and the API key if needed}
\usage{
set_config()
}
\description{
This configuration wizard will help you to configure the LibreTranslate instance
you want to use. It also support storing the API key and exporting the configuration
in a .json file that can be read later with \code{import_config()} to avoid running this
command every time you initiate R.
}
\examples{
\dontrun{
set_config()
}
}
