% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate strings using LibreTranslate selected instance}
\usage{
translate(q, from = "auto", to)
}
\arguments{
\item{q}{The string you want to translate}

\item{from}{The origin language of the text. If you leave it empty, it will be set to "auto".}

\item{to}{The destination language you want the string to be translated to.}
}
\description{
Request as many details as you want (revenue, cast, original language, ...)
from media and add them to an existing dataframe.
}
\examples{
\dontrun{
translate(q = "hello, nice to meet you", from = "en", to = "es")
}
}
