% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_file}
\alias{translate_file}
\title{Translate files using LibreTranslate selected instance}
\usage{
translate_file(file, from = "auto", to)
}
\arguments{
\item{file}{The path to the file you want to translate (absolute or relative).}

\item{from}{The origin language of the text. If you leave it empty, it will be set to "auto".}

\item{to}{The destination language you want the text to be translated to.}
}
\description{
With this function, you can translate an external file.
Supported formats are .txt, .odt, .odp, .docx, .pptx, .epub and .html
}
\examples{
\dontrun{
translate_file <- function(file, from = "auto", to)
}
}
